bugfix in AO

+++ AO_Default Web Server/Support/JS_library/ByteArray_convertor.js
@@ -1,8 +1,4 @@
 function convertToByteArray(arr1, arr2) {
-	//if (arr1.length !== arr2.length) {
-	//    throw new Error("Both arrays must have the same length.");
-
-	//  }// Merge both DBL arrays into a single array
 	const rows = arr1.length;
 	// Create a Float32Array with two columns per row.
 	const floatArray = new Float32Array(rows * 2);
@@ -107,7 +103,7 @@ function parsePythonBytesLiteral(pyStr) {
 	
 	// Create a Float32Array view over the same buffer.
 	// (Assumes the environment uses little-endian storage.)
-	const floatArray = new Float32Array(
+	const floatArray = new Int32Array(
 	  byteArray.buffer, 
 	  byteArray.byteOffset, 
 	  byteArray.length / 4
@@ -122,7 +118,7 @@ function parsePythonBytesLiteral(pyStr) {
 	}
 	
 	// Combine the two channels with the separator in between.
-	const returnFloatArray = new Float32Array(channel0.concat([separator]).concat(channel1));
+	const returnFloatArray = new Int32Array(channel0.concat([separator]).concat(channel1));
 	// Convert the combined array into a typed Float32Array.
 	return (returnFloatArray);
   }
